refactor: Store内の型情報を一箇所に集約
