+++ src/features/template-search/store/types.ts
@@ -1,0 +1,7 @@
+export type FavoriteTemplateList = {
+    ids: string[]
+}
+
+export type TemplateSearchSettings = {
+    featureUsage: boolean
+}
